refactor: LayerPanel にLayerPanelContentsコンポ―ネントを統合 #7
move:LayerColumからLayerPanelフォルダに移動
